Changed ibc testbitset finalizers_count param type to uint32_t, recompiled contract

+++ unittests/test-contracts/savanna/ibc/ibc.cpp
@@ -169,7 +169,7 @@ ACTION ibc::checkproof(const proof& proof){
 
 }
 
-ACTION ibc::testbitset(const std::string bitset_string, const std::vector<uint8_t> bitset_vector, const size_t finalizers_count){
+ACTION ibc::testbitset(const std::string bitset_string, const std::vector<uint8_t> bitset_vector, const uint32_t finalizers_count){
     savanna::bitset b(finalizers_count, bitset_vector);
 
     check(b.to_string() == bitset_string, "bitset mismatch");
